Bug fix and improved responsibility

+++ be/courses.js
@@ -232,7 +232,7 @@ async function getWords(req, validate) {
 			INNER JOIN words AS wo ON gr.word = wo.id
 			LEFT JOIN word_state AS ws ON ws.word_group = gr.id AND ws."user" = ?
 			LEFT JOIN (
-	            SELECT MAX(changed), count(CASE when state = 'known' then 1 end) AS known_times, word_group FROM word_state
+	            SELECT MAX(changed), count(CASE when state = 'known' then 1 end)::INT AS known_times, word_group FROM word_state
 	            WHERE "user" = ? GROUP BY ("user", "word_group")
 	        ) states ON states.word_group = gr.id
         WHERE course_node = ? AND
